Refactored Backpropagation into its own Class
